Backend Refactor:  - Reorganized backend using express router  - Added origin and methods specification in CORS for security
