+++ backend/.env.example
@@ -1,2 +1,3 @@
 PORT=8080
+FRONTEND_HOST="http://localhost:3000"
 SESSION_SECRET="YOUR_SECRET_HERE"
